Reboot if the Linux kernel panics
Prior to this commit, Linux would hang indefinitely. If enabled and
started, a hardware watchdog would reboot in theory, but that's not
guaranteed in all setups.  Since no good can come from a panic, pause
for a little bit (10 seconds) and reboot. The pause is good for
throttling the continuous panic case, but the hope would be that panics
are rare occurances in the field.