Enable the Raspberry Pi Touchscreen by default
This makes it possible to support Scenic using an official Nerves
system on the popular Raspberry Pi 7" Touchscreen.

This appears harmless to non-touchscreen users, but I left a comment in
the config.txt in the hopes that if any negative effects were seen by
the community that we can at least know about them.