Update partitioning to support tryboot for automatic fallback
!IMPORTANT! This is a one way upgrade. Going back to the old partitioning
requires manually reflashing of the RPi's storage.

This updates the `fwup` configuration to support the Raspberry Pi
tryboot feature so that it's possible to recover from messed up Linux
kernels and OTP boot failures.

This works by telling the system to reboot into the new firmware slot
one time. If the firmware slot is good, then some code (Nerves.Runtime
can do this) needs to call `Nerves.Runtime.validate_firmware`. That will
cause subsequent reboots to use the new slot.

The theory is that the app won't validate the firmware unless it's good,
so that if anything crashes, the previously working firmware will be
used.

Even though the partitioning looks different, it's possible to upgrade
non-tryboot Nerves Raspberry Pi firmware without losing application
data. This is supported by lining up partitions so that they're close to
the same places. Downgrading is not possible, though. Also, there's a window
where cancelling the firmware update can brick the Pi due to the rootfs
partition needing to be shifted, so be careful. There's a print in the
fwup config to remind about this.